Sport Scores - Observer Pattern Application
The idea is to emulate a service that supplies the results of college
sporting events to interested clients, similar to an RSS feed.  Each
time the service receives the score of an event, it notifies each of
its clients.  A client can then do what it likes with this information.